transformer/requirements.txt
flask==3.0.3
gunicorn==23.0.0
flask-flatpages==0.8.2
transformers==4.57.0
pillow==10.4.0
numpy==1.26.4
requests==2.32.5
scipy==1.16.2
cloudinary==1.44.1
flask-cors==6.0.1
logging==0.4.9.6
transformers==4.57.0
torch==2.8.0
omegaconf==2.3.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
trimesh==4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
gradio
xatlas==0.0.9
moderngl==5.10.0